refactor: rename vim_parser -> buf_parser
